requirements.txt
Flask
requests
botframework-connector
botbuilder-schema
botbuilder-core
botbuilder-dialogs
datatypes-date-time
msrest

